Integrate: [ 33260] I've been maintaining Net::Ping for a while now.
[ 33295]
Make sure we only find the macro we were looking for,
not something that only starts with the same string.

[ 33303]
Jerry D. Hedden now maintains Thread::Semaphore on CPAN

[ 33352]
Subject: [patch] Porting/expand-macros.pl gets 'indent'ing
From: Jim Cromie <jim.cromie@gmail.com>
Message-ID: <47AB9C4A.3080107@gmail.com>
Date: Thu, 07 Feb 2008 17:03:22 -0700

[ 33353]
Take advantage of the fact that we can use indent as a stdin/stdout
filter to reduce its workload (and ours) by only sending it the 3 or
so lines that we are interested in printing, not the preceding
bucket loads.

[ 33456]
Subject: checkAUTHORS.pl update for me
From: Michael G Schwern <schwern@pobox.com>
Date: Sun, 09 Mar 2008 00:35:48 +0100
Message-ID: <47D322D4.8090508@pobox.com>
p4raw-link: @33456 on //depot/perl: 8ed0547917b5366519f80a9298ced6aebb27aeef
p4raw-link: @33353 on //depot/perl: 27c6397cdef882675fd24e716944ef4ce35abe7b
p4raw-link: @33352 on //depot/perl: a93e78e3a0993367922540f294652a138d1d7225
p4raw-link: @33303 on //depot/perl: 9f4ee62979dc352c40be9112d4effb2365866c3a
p4raw-link: @33295 on //depot/perl: e25a7dc23f32d0dd4f1c4c8645877a39bdb9d4b7
p4raw-link: @33260 on //depot/perl: 7e8c2d18d3c652cc69ad75e73b2b8f2ab220bfbd

p4raw-id: //depot/maint-5.10/perl@33617
p4raw-integrated: from //depot/perl@33616 'copy in'
	Porting/checkAUTHORS.pl (@31910..)
p4raw-integrated: from //depot/perl@33295 'ignore'
	Porting/expand-macro.pl (@32833..)
p4raw-integrated: from //depot/perl@33260 'merge in'
	Porting/Maintainers.pl (@33090..)

+++ Porting/expand-macro.pl
@@ -1,16 +1,37 @@
 #!perl -w
 use strict;
 
+use Getopt::Std;
+
 use vars qw($trysource $tryout $sentinel);
 $trysource = "try.c";
 $tryout = "try.i";
 
+getopts('fF:ekvI:', \my %opt) or usage();
+
+sub usage {
+    die<<EO_HELP;
+@_;
+usage: $0 [options] <macro-name> [headers]
+options:
+    -f		use 'indent' to format output
+    -F	<tool>	use <tool> to format output  (instead of -f)
+    -e		erase try.[ic] instead of failing when theyre present (errdetect)
+    -k		keep them after generating (for handy inspection)
+    -v		verbose
+    -I <indent-opts>	passed into indent
+EO_HELP
+}
+
 my $macro = shift;
-die "$0 macro [headers]" unless defined $macro;
+usage "missing <macro-name>" unless defined $macro;
 
 $sentinel = "$macro expands to";
 
+usage "-f and -F <tool> are exclusive\n" if $opt{f} and $opt{F};
+
 foreach($trysource, $tryout) {
+    unlink $_ if $opt{e};
     die "You already have a $_" if -e $_;
 }
 
@@ -23,16 +44,20 @@ if (!@ARGV) {
 
 my $args = '';
 
+my $found_macro;
 while (<>) {
-    next unless /^#\s*define\s+$macro/;
+    next unless /^#\s*define\s+$macro\b/;
     my ($def_args) = /^#\s*define\s+$macro\(([^)]*)\)/;
     if (defined $def_args) {
 	my @args = split ',', $def_args;
+	print "# macro: $macro args: @args in $_\n" if $opt{v};
 	my $argname = "A0";
 	$args = '(' . join (', ', map {$argname++} 1..@args) . ')';
     }
+    $found_macro++;
     last;
 }
+die "$macro not found\n" unless $found_macro;
 
 open my $out, '>', $trysource or die "Can't open $trysource: $!";
 
@@ -45,14 +70,38 @@ EOF
 
 close $out or die "Can't close $trysource: $!";
 
+print "doing: make $tryout\n" if $opt{v};
 system "make $tryout" and die;
+
+# if user wants 'indent' formatting ..
+my $out_fh;
+
+if ($opt{f} || $opt{F}) {
+    # a: indent is a well behaved filter when given 0 arguments, reading from
+    #    stdin and writing to stdout
+    # b: all our braces should be balanced, indented back to column 0, in the
+    #    headers, hence everything before our #line directive can be ignored
+    #
+    # We can take advantage of this to reduce the work to indent.
+
+    my $indent_command = $opt{f} ? 'indent' : $opt{F};
+
+    if (defined $opt{I}) {
+	$indent_command .= " $opt{I}";
+    }
+    open $out_fh, '|-', $indent_command or die $?;
+} else {
+    $out_fh = \*STDOUT;
+}
 
 open my $fh, '<', $tryout or die "Can't open $tryout: $!";
 
 while (<$fh>) {
-    print if /$sentinel/o .. 1;
+    print $out_fh $_ if /$sentinel/o .. 1;
 }
 
-foreach($trysource, $tryout) {
-    die "Can't unlink $_" unless unlink $_;
+unless ($opt{k}) {
+    foreach($trysource, $tryout) {
+	die "Can't unlink $_" unless unlink $_;
+    }
 }
